apollo_infra: move server connection eviction test to designated module
